+++ packages/components/stencil-staging.config.ts
@@ -1,5 +1,6 @@
 import { Config } from '@stencil/core';
 import { postcss } from '@stencil/postcss';
+import postcssCustomMedia from 'postcss-custom-media';
 import { inlineSvg } from 'stencil-inline-svg';
 
 export const config: Config = {
@@ -9,7 +10,7 @@ export const config: Config = {
   plugins: [
     inlineSvg(),
     postcss({
-      plugins: [], // TODO
+      plugins: [postcssCustomMedia()],
     }),
   ],
   outputTargets: [