+++ packages/storybook-vue/stories/components/text-field/ScaleTextField.vue
@@ -18,8 +18,10 @@
     :counter="counter"
     :size="size"
     :value="value"
+    :aria-detailed-id="ariaDetailedId"
     :input-id="inputId"
     :input-autofocus="inputAutofocus"
+    :input-autocomplete="inputAutocomplete"
     :experimental-controlled="experimentalControlled"
     @scaleInput="scaleInput"
     @scaleChange="scaleChange"
@@ -54,8 +56,10 @@ export default {
     counter: { type: Boolean },
     size: { type: String },
     value: { type: String },
+    ariaDetailedId: { type: String },
     inputId: { type: String },
     inputAutofocus: { type: Boolean },
+    inputAutocomplete: { type: String },
     experimentalControlled: { type: Boolean },
     styles: { type: String },
   },